backend/auth/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
email-validator>=2.0.0
stripe>=7.0.0
cryptography>=42.0.0
